cloud_functions/view_group/requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-datastore
jinja2
